append local reference table placement to dropStmt if needed
If one of the range table entries is indicating reference table
whose local placement referencing to a local table, PostgreSQL
wouldn't be aware of that we are dropping the dependent table too
and hence would error out (if drop command is not cascading). To
solve this problem, we basically append the local placements of
reference tables to the range table entries if they are referenced
by a local table.