Fix assertion error when rolling back to savepoint
Fixes #3622

It was possible to get an assertion error, if a DML command was
cancelled that opened a connection and then "ROLLBACK TO SAVEPOINT" was
used to continue the transaction. The reason for this was that canceling
the transaction might leave the `claimedExclusively` flag on for (some
of) it's connections.

This caused an assertion failure because `CanUseExistingConnection`
would return false and a new connection would be opened, and then there
would be two connections doing DML for the same placement. Which is
disallowed. That this situation caused an assertion failure instead of
an error, means that without asserts this could possibly result in some
visibility bugs, similar to the ones described
https://github.com/citusdata/citus/issues/3867

The fix simply "unclaims" all connections after "ROLLBACK TO SAVEPOINT"
is done.

This specific issue also highlights some other issues:
1. Just like #3867, it shows that the `CanUseExistingConnection` should
   be improved to catch these type of errors.
2. I think we should make these DML and DDL asserts normal errors.
   This way we will not return incorrect data in case of bugs, but instead
   error out:
   https://github.com/citusdata/citus/blob/master/src/backend/distributed/connection/placement_connection.c#L418-L419
   This code is complex and it's quite possible we missed some other
   edge cases.
3. "ROLLBACK TO SAVEPOINT" in plain Postgres undos the locks that the
   rolled back statements took. We do not undo our hadDML and hadDDL
   flags, when rolling back to a savepoint. This could result in some
   queries not being allowed on Citus that would actually be fine to
   execute. Changing this would require us to keep track for each
   savepoint which placement connections had DDL/DML executed at that
   point.